python/requirements.txt
fastapi
uvicorn
pydantic
google-cloud-bigquery-storage
pyarrow
adbc-driver-postgresql
adbc-driver-bigquery
pyyaml
pyarrow-bigquery
